models/cluster_models/requirements.txt
aiosqlite==0.20.0
apscheduler==3.11.0
FastAPI==0.115.6
httpx==0.26.0
httpx[cli]
httpx[http2]
imblearn==0.0
joblib==1.4.2
matplotlib==3.8.4
numpy==1.23.4
pandas==2.0.3
pandas_ta==0.3.14b0
plotly==5.24.1
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==20.8
pytz==2024.2
requests==2.26.0
scikit-learn==1.6.0
scipy==1.10.1
sqlalchemy==2.0.36
sqlalchemy-utc==0.14.0
sqlmodel==0.0.22
tqdm==4.62.3
uvicorn==0.32.1
yfinance==0.2.18